WIP: Rename package name from fakeserver to mockcoreserver
