+++ test/e2e/pkg/mockcoreserver/call.go
@@ -1,4 +1,4 @@
-package fakeserver
+package mockcoreserver
 
 import (
 	"io"
