+++ test/e2e/pkg/mockcoreserver/expect.go
@@ -1,4 +1,4 @@
-package fakeserver
+package mockcoreserver
 
 import (
 	"bytes"
