+++ test/e2e/pkg/mockcoreserver/handler.go
@@ -1,4 +1,4 @@
-package fakeserver
+package mockcoreserver
 
 import (
 	"net/http"
