+++ test/e2e/pkg/mockcoreserver/server.go
@@ -1,4 +1,4 @@
-package fakeserver
+package mockcoreserver
 
 import (
 	"context"
